Made Pawns visible while selecting
The pawns were geting hidden behind other sprites  previousely.
I have updated the z-index of all the pawns belonging to the selcted player so that they are drawn above other sprites during selection of pawn for movement. After the movement is complete the z-index is reset back to 0 so that teh next player pawns can be drawn in front